% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{first}
\alias{first}
\title{first}
\usage{
first(seq)
}
\arguments{
\item{seq}{List or vector}
}
\value{
The first element in a list or vector
}
\description{
Returns the first element of a vector or list. NA if out of bounds.
}
