% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.R
\name{rest}
\alias{rest}
\title{rest}
\usage{
rest(seq)
}
\arguments{
\item{seq}{List or vector}
}
\value{
Everything in the list or vector except the first element
}
\description{
Returns all but the first element of a vector or list. NA if out of bounds.
}
\examples{
rest(c(1,2,3))
}
